Update README with SIP link and build instructions
The SIP process is detailed in a spreadsheet on Google Drive. It contains an overall product backlog (i.e. user stories), 2 sprint backlogs, a task list that aims to accomplish the user stories, a time log detailing the time spent on various tasks, and a glossary. The spreadsheet will be updated as development takes place.