Cppcheck, Cpplint results and example image
User Story ID: All; Task ID: All. Cppcheck and cpplint results screenshots are included. Note that only one error is present in the cpplint output and it has to deal with multithreading performance improvements, which is not a consideration for this implementation at this time. Also included is an example image output from the module compared to the source image.